pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refunc"
version = "0.1.0"
description = "A comprehensive ML utilities toolkit with advanced logging, performance monitoring, and file handling."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kennedy DS", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "psutil>=5.8.0",
    "colorama>=0.4.4",
    "tqdm>=4.62.0",
    "pyyaml>=5.4.0",
    "python-dateutil>=2.8.0",
    "typing-extensions>=3.10.0;python_version<'3.8'",
    "pyarrow>=5.0.0",
    "openpyxl>=3.0.0",
    "h5py>=3.0.0",
    "joblib>=1.1.0",
]

[project.optional-dependencies]
gpu = ["GPUtil>=1.4.0", "py3nvml>=0.2.0"]
dev = [
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
    "pre-commit>=2.15.0",
]
test = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "pytest-mock>=3.6.0",
    "pytest-benchmark>=3.4.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "jupyter>=1.0.0",
    "nbsphinx>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/kennedym-ds/refunc"
Repository = "https://github.com/kennedym-ds/refunc"
Issues = "https://github.com/kennedym-ds/refunc/issues"
Documentation = "https://github.com/kennedym-ds/refunc/blob/main/README.md"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["refunc"]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=refunc --cov-report=term-missing"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU (deselect with '-m \"not gpu\"')",
]
